refactor(dice): update type system and add initial tests
- Change DiceResult parameters from short to int for consistency
- Add basic stub implementation in DiceRoller
- Set up initial test structure
- Update code coverage requirements to 95%

+++ src/main/java/com/sat2/dice/DiceResult.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 @Value
 public class DiceResult {
-    short diceCount;
-    short diceSides;
+    int diceCount;
+    int diceSides;
     List<Integer> results;
     int total;
 
-    public DiceResult(short diceCount, short diceSides, List<Integer> results) {
+    public DiceResult(int diceCount, int diceSides, List<Integer> results) {
         this.diceCount = diceCount;
         this.diceSides = diceSides;
         this.results = List.copyOf(results);
